파이썬프로그래밍2주차

+++ inTocm.py
@@ -1,3 +1,3 @@
-a = int(input("Inch 값을 입력하세요 : "))
+a = float(input("Inch 값을 입력하세요 : "))
 result = a * 2.54
 print("=", result, 'cm')
